Adds user restore functionality
Enables administrators to restore soft-deleted users.

Implements a new `restore` method in the `UserController` that
allows restoring a soft-deleted user by ID.  It includes
authorization checks to ensure only admins can perform this action.

Adds a corresponding policy rule to the `UserPolicy` to
control access to the `restore` functionality based on user role.

Adds a seeder for the Profile model

Fixes #42