Enables authorization for task management
This commit enables authorization checks in the TaskController for index, store, and show methods.
It also updates the TaskFactory to use the fully qualified namespace for User and Lesson models.
Additionally, it updates the TaskSeeder to create a specific number of tasks.
It also updates the create_categories_table migration to use static function